till 1380 delete by id ,,,,tested company working fine ----do test all APIs on review tommorow mornig for sure

+++ src/main/java/com/projects/firstjobapp/job/JobRepository.java
@@ -2,7 +2,5 @@ package com.projects.firstjobapp.job;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobRepository extends JpaRepository<Job, Integer> {
-
-
+public interface JobRepository extends JpaRepository<Job, Long> {
 }
